+++ lib/global/course_progress_tracker.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:wordy/Utility/locator/service_locator.dart';
 import 'package:wordy/const/enums.dart';
 import 'package:wordy/domain/models/active_course.dart';
 
@@ -10,16 +11,19 @@ class CourseProgressTracker extends ChangeNotifier {
 
   void setQuizType(QuizType quizType) {
     this.quizType = quizType;
+    locator<CourseProgressTracker>().quizType = quizType;
     notifyListeners();
   }
 
   void setBeforeQuiz(ActiveCourse? beforeQuiz) {
     this.beforeQuiz = beforeQuiz;
+    locator<CourseProgressTracker>().beforeQuiz = beforeQuiz;
     notifyListeners();
   }
 
   void setAfterQuiz(ActiveCourse? afterQuiz) {
     this.afterQuiz = afterQuiz;
+    locator<CourseProgressTracker>().afterQuiz = afterQuiz;
     notifyListeners();
   }
 }
