+++ lib/presentation/Bloc/quiz/quiz_bloc.dart
@@ -71,8 +71,10 @@ class QuizBloc extends Bloc<QuizEvent, QuizState> {
       }
 
       if (questionsData.isData) {
+        print('QUIZ TYPE IS  ' + quizType.toString());
         courseName = courseNameData.data!.course.name;
         questions = questionsData.data!;
+
         emit(QuizQuestionState(
             question: questions[currentQuestionIndex],
             selectedIndex: null,
